pyproject.toml
[tool.poetry]
name = "airflow-project"
version = "0.1.0"
description = ""
authors = ["Rafael Panegassi <rm.panegassi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
apache-airflow = "^2.10.1"
loguru = "^0.7.2"
streamlit = "^1.38.0"
requests = "^2.32.3"
pydantic = "^2.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
